+++ worker/src/DepUsrSCTP.cpp
@@ -2,6 +2,9 @@
 // #define MS_LOG_DEV_LEVEL 3
 
 #include "DepUsrSCTP.hpp"
+#ifdef MS_LIBURING_SUPPORTED
+#include "DepLibUring.hpp"
+#endif
 #include "DepLibUV.hpp"
 #include "Logger.hpp"
 #include <usrsctp.h>
@@ -249,7 +252,21 @@ void DepUsrSCTP::Checker::OnTimer(TimerHandle* /*timer*/)
 	auto nowMs          = DepLibUV::GetTimeMs();
 	const int elapsedMs = this->lastCalledAtMs ? static_cast<int>(nowMs - this->lastCalledAtMs) : 0;
 
+#ifdef MS_LIBURING_SUPPORTED
+	// Activate liburing usage.
+	// 'usrsctp_handle_timers()' will synchronously call the send/recv
+	// callbacks for the pending data. If there are multiple messages to be
+	// sent over the network then we will send those messages within a single
+	// system call.
+	DepLibUring::SetActive();
+#endif
+
 	usrsctp_handle_timers(elapsedMs);
+
+#ifdef MS_LIBURING_SUPPORTED
+	// Submit all prepared submission entries.
+	DepLibUring::Submit();
+#endif
 
 	this->lastCalledAtMs = nowMs;
 }